+++ src/main/java/open/commons/spring/elastic/service/AbstractElasticsearchService.java
@@ -26,6 +26,15 @@
 
 package open.commons.spring.elastic.service;
 
+import java.io.BufferedReader;
+import java.io.File;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.Reader;
+import java.io.StringReader;
+import java.io.Writer;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 import java.util.function.Supplier;
@@ -48,6 +57,8 @@ import org.springframework.data.elasticsearch.core.query.IndexQueryBuilder;
 import org.springframework.data.elasticsearch.core.query.Query;
 
 import open.commons.core.Result;
+import open.commons.core.utils.IOUtils;
+import open.commons.spring.elastic.utils.RestApiUtils;
 import open.commons.spring.web.mvc.service.AbstractComponent;
 
 /**
@@ -85,6 +96,46 @@ public class AbstractElasticsearchService extends AbstractComponent {
         this.esClientConfig = esClientConfig;
         this.highLevelClient = RestClients.create(this.esClientConfig).rest();
         this.lowLevelClient = RestClients.create(this.esClientConfig).lowLevelRest();
+    }
+
+    /**
+     * Elasticsearch JSON 으로 저장된 파일을 한번에 저장합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2022. 9. 15.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param url
+     *            Elasticsearch API 'Bulk Index' URL.<br>
+     *            <ul>
+     *            <li>포맷: {scheme}://{host}:{port}/{index}/_index
+     *            <li>예시: http://192.168.0.10:80/test-index/_index
+     *            </ul>
+     * @param filepath
+     *            Elasticsearch JSON bulk 파일 절대경로
+     * @throws IOException
+     * @throws InterruptedException
+     *
+     * @since 2022. 9. 15.
+     * @version _._._
+     * @author Park Jun-Hong (parkjunhong77@gmail.com)
+     */
+    public void bulkIndex(String url, String filepath) throws IOException, InterruptedException {
+        String[] cmdarr = { "curl", "-X", "POST", url, "-H", "'Content-Type: application/x-ndjson'", "--data-binary", "@" + filepath };
+        Process bulkProc = Runtime.getRuntime().exec(cmdarr);
+        InputStream in = bulkProc.getInputStream();
+        while (in.read() != -1) {
+            ;
+        }
+        bulkProc.waitFor();
+
+        int exitCode = bulkProc.exitValue();
+        if (exitCode != 0) {
+            logger.warn("Elasticsearch 'bulk' API 실행을 비정상적({})으로 종료되었습니다. command={}", exitCode, Arrays.toString(cmdarr));
+        }
     }
 
     /**
@@ -145,6 +196,73 @@ public class AbstractElasticsearchService extends AbstractComponent {
         };
 
         return action;
+    }
+
+    /**
+     * 데이터를 Elasticsearch Bulk JSON 형태로 변환하여 제공합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2022. 9. 15.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param <T>
+     * @param data
+     * @return
+     *
+     * @since 2022. 9. 15.
+     * @version 0.2.0
+     * @author Park Jun-Hong (parkjunhong77@gmail.com)
+     */
+    public <T> String createBulkJSONString(@NotNull List<T> data) {
+        return data.parallelStream()
+                // 단일 객체를 문자열로 변환
+                .map(d -> RestApiUtils.toIndexBulkString(d))
+                // 'null' 확인
+                .filter(s -> s != null)
+                // 하나의 문자열로 통합
+                .collect(Collectors.joining());
+    }
+
+    /**
+     * 데이터를 Elasticsearch JSON 문자열로 변환하여 저장한 후, 저장한 파일정보를 제공합니다. <br>
+     * 
+     * <pre>
+     * [개정이력]
+     *      날짜    	| 작성자	|	내용
+     * ------------------------------------------
+     * 2022. 9. 15.		박준홍			최초 작성
+     * </pre>
+     *
+     * @param <T>
+     *            데이터 타입
+     * @param data
+     *            데이터
+     * @param tempfilePrefix
+     *            임시파일 접두어
+     * @param tempfileSuffix
+     *            임시파일 접미어
+     * @return 임시파일 절대 경로
+     * @throws IOException
+     *             파일 생성시 오류가 발생한 경우
+     *
+     * @since 2022. 9. 15.
+     * @version 0.2.0
+     * @author Park Jun-Hong (parkjunhong77@gmail.com)
+     */
+    public <T> String createBulkJSONTempFile(List<T> data, String tempfilePrefix, String tempfileSuffix) throws IOException {
+        // #1. bulk JSON 문자열 생성
+        String bulkJSONStr = createBulkJSONString(data);
+        Reader reader = new BufferedReader(new StringReader(bulkJSONStr));
+
+        // #2. bulk JSON 파일 생성
+        File tmpfile = File.createTempFile(tempfilePrefix, tempfileSuffix);
+        Writer writer = new FileWriter(tmpfile);
+
+        IOUtils.transfer(reader, true, writer, true, IOUtils.BUFFER_SIZE_1MB);
+        return tmpfile.getAbsolutePath();
     }
 
     /**